- User query refactored
