+++ ui/src/lib/gql/index.ts
@@ -1,3 +1,3 @@
-export * from "./queries";
 export * from "./provider/";
+export * from "./user";
 export * from "./workspace";
