--- ui/src/lib/gql/queries/index.ts
@@ -1,1 +1,0 @@
-export * from "./user";
