+++ ui/src/lib/gql/user/index.ts
@@ -1,0 +1,1 @@
+export * from "./hook";
