+++ ui/src/lib/gql/user/queries.ts
@@ -1,0 +1,12 @@
+import { graphql } from "../__gen__";
+
+graphql(`
+  query GetMe {
+    me {
+      id
+      name
+      email
+      myWorkspaceId
+    }
+  }
+`);
