+++ ui/src/types/user.ts
@@ -1,7 +1,15 @@
+import { ApiResponse } from "./api";
+
+export type GetMe = {
+  me: User | undefined | null;
+  isLoading: boolean;
+} & ApiResponse;
+
 export type User = {
   id: string;
   name: string;
   email: string;
+  myWorkspaceId: string;
   lang?: string;
   theme?: string;
   // workspace?: string;
